fix(): fix use brick with template
