+++ packages/brick-kit/src/core/CustomTemplates/expandCustomTemplate.ts
@@ -1,16 +1,11 @@
 import {
-  BrickConf,
   BrickConfInTemplate,
   CustomTemplateProxySlot,
   PluginRuntimeContext,
-  RefForProxy,
   RuntimeBrickConf,
   SlotsConfOfBricks,
   CustomTemplate,
 } from "@next-core/brick-types";
-import { hasOwnProperty } from "@next-core/brick-utils";
-import { clamp } from "lodash";
-import { preprocessTransformProperties } from "../../transformProperties";
 import { RuntimeBrick } from "../BrickNode";
 import {
   MergeBase,
@@ -18,26 +13,18 @@ import {
   RuntimeCustomTemplateProxy,
   RuntimeCustomTemplateProxyProperties,
 } from "./internalInterfaces";
-import {
-  isBasicProperty,
-  isMergeableProperty,
-  isTransformableProperty,
-  isVariableProperty,
-} from "./assertions";
+import { isMergeableProperty, isVariableProperty } from "./assertions";
 import { collectRefsInTemplate } from "./collectRefsInTemplate";
 import {
   customTemplateRegistry,
   RuntimeBrickConfWithTplSymbols,
-  symbolForComputedPropsFromProxy,
-  symbolForRefForProxy,
-  symbolForTplContextId,
 } from "./constants";
-import { propertyMergeAll } from "./propertyMerge";
 import { collectMergeBases } from "./collectMergeBases";
 import { CustomTemplateContext } from "./CustomTemplateContext";
 import { setupUseBrickInTemplate } from "./setupUseBrickInTemplate";
+import { setupTemplateProxy } from "./setupTemplateProxy";
 
-interface ProxyContext {
+export interface ProxyContext {
   reversedProxies: ReversedProxies;
   templateProperties: Record<string, unknown>;
   externalSlots: SlotsConfOfBricks;
@@ -235,15 +222,6 @@ function expandBrickInTemplate(
     slots: slotsInTemplate,
     ...restBrickConfInTemplate
   } = brickConfInTemplate;
-  const {
-    reversedProxies,
-    templateProperties,
-    externalSlots,
-    templateContextId,
-    proxyBrick,
-  } = proxyContext;
-  const computedPropsFromProxy: Record<string, unknown> = {};
-  let refForProxy: RefForProxy;
 
   const slots: SlotsConfOfBricks = Object.fromEntries(
     Object.entries(slotsInTemplate ?? {}).map(([slotName, slotConf]) => [
@@ -257,114 +235,11 @@ function expandBrickInTemplate(
     ])
   );
 
-  setupUseBrickInTemplate(brickConfInTemplate.properties, templateContextId);
-
-  if (ref) {
-    refForProxy = {};
-    proxyBrick.proxyRefs.set(ref, refForProxy);
-
-    // Reversed proxies are used for expand storyboard before rendering page.
-    if (reversedProxies.properties.has(ref)) {
-      Object.assign(
-        computedPropsFromProxy,
-        Object.fromEntries(
-          reversedProxies.properties
-            .get(ref)
-            .flatMap((item) => {
-              // `propValue` is computed.
-              const propValue = templateProperties?.[item.$$reversedRef];
-              if (isTransformableProperty(item)) {
-                return Object.entries(
-                  preprocessTransformProperties(
-                    {
-                      [item.$$reversedRef]: propValue,
-                    },
-                    item.refTransform
-                  )
-                );
-              }
-              if (isBasicProperty(item)) {
-                return [[item.refProperty, propValue]];
-              }
-              // Ignore Variable properties.
-              // And mergeable properties are processed later.
-              return [];
-            })
-            .filter((item) => item[1] !== undefined)
-        )
-      );
-    }
-
-    // Brick properties can be merged multiple times.
-    if (reversedProxies.mergeBases.has(ref)) {
-      Object.assign(
-        computedPropsFromProxy,
-        Object.fromEntries(
-          Array.from(reversedProxies.mergeBases.get(ref).entries())
-            .map(([mergeProperty, mergeBase]) => [
-              mergeProperty,
-              propertyMergeAll(mergeBase, templateProperties ?? {}),
-            ])
-            .filter((item) => item[1] !== undefined)
-        )
-      );
-    }
-
-    // Use an approach like template-literal's quasis:
-    // `quasi0${0}quais1${1}quasi2...`
-    // Every quasi (indexed by `refPosition`) can be slotted with multiple bricks.
-    const quasisMap = new Map<string, BrickConf[][]>();
-
-    if (reversedProxies.slots.has(ref)) {
-      for (const item of reversedProxies.slots.get(ref)) {
-        if (!quasisMap.has(item.refSlot)) {
-          const quasis: BrickConf[][] = [];
-          // The size of quasis should be the existed slotted bricks' size plus one.
-          const size = hasOwnProperty(slots, item.refSlot)
-            ? slots[item.refSlot].bricks.length + 1
-            : 1;
-          for (let i = 0; i < size; i += 1) {
-            quasis.push([]);
-          }
-          quasisMap.set(item.refSlot, quasis);
-        }
-        const expandableSlot = quasisMap.get(item.refSlot);
-        const refPosition = item.refPosition ?? -1;
-        expandableSlot[
-          clamp(
-            refPosition < 0 ? expandableSlot.length + refPosition : refPosition,
-            0,
-            expandableSlot.length - 1
-          )
-        ].push(...(externalSlots?.[item.$$reversedRef]?.bricks ?? []));
-      }
-    }
-
-    for (const [slotName, quasis] of quasisMap.entries()) {
-      if (!hasOwnProperty(slots, slotName)) {
-        slots[slotName] = {
-          type: "bricks",
-          bricks: [],
-        };
-      }
-      const slotConf = slots[slotName];
-      slotConf.bricks = quasis.flatMap((bricks, index) =>
-        index < slotConf.bricks.length
-          ? bricks.concat(slotConf.bricks[index])
-          : bricks
-      );
-
-      if (slotConf.bricks.length === 0) {
-        delete slots[slotName];
-      }
-    }
-  }
+  setupUseBrickInTemplate(brickConfInTemplate.properties, proxyContext);
 
   return {
     ...restBrickConfInTemplate,
     slots,
-    [symbolForComputedPropsFromProxy]: computedPropsFromProxy,
-    [symbolForRefForProxy]: refForProxy,
-    [symbolForTplContextId]: templateContextId,
+    ...setupTemplateProxy(proxyContext, ref, slots),
   };
 }
